requirements.txt
discord.py==1.7.2
sentry-sdk==1.1.0
